requirements.txt
Django==2.1
djangorestframework==3.8.2
mysqlclient==1.3.13
requests==2.19.1
inflection==0.3.1